+++ src/main/java/calculator/StringAddCalculator.java
@@ -5,7 +5,6 @@ import java.util.regex.Pattern;
 
 public class StringAddCalculator {
 
-    public static final String DEFAULT_DELIMITER = ",|:";
     public static final String CUSTOM_EXPRESSION = "//(.)\n(.*)";
     public static final Pattern CUSTOM_PATTERN = Pattern.compile(CUSTOM_EXPRESSION);
 
@@ -15,7 +14,7 @@ public class StringAddCalculator {
     public static int splitAndSum(String text) {
         if (hasNotText(text))
             return 0;
-        if (hasCustomDelimiter(text)) {
+        if (Pattern.matches(CUSTOM_EXPRESSION, text)) {
             return calculateSum(makeCustomSplitDto(text));
         }
         return calculateSum(SplitDto.of(text));
@@ -38,16 +37,6 @@ public class StringAddCalculator {
 
     private static boolean isEmptyString(String text) {
         return "".equals(text);
-    }
-
-    private static SplitDto makeCustomSplitDto(String text) {
-        Matcher m = CUSTOM_PATTERN.matcher(text);
-        if (m.find()) {
-            String customDelimiter = m.group(1);
-            String extractedText = m.group(2);
-            return SplitDto.of(extractedText, customDelimiter);
-        }
-        throw new RuntimeException();
     }
 
     private static int calculateSum(SplitDto dto) {
@@ -75,31 +64,13 @@ public class StringAddCalculator {
         }
     }
 
-    private static boolean hasCustomDelimiter(String text) {
-        return Pattern.matches(CUSTOM_EXPRESSION, text);
+    private static SplitDto makeCustomSplitDto(String text) {
+        Matcher matcher = CUSTOM_PATTERN.matcher(text);
+        if (matcher.find()) {
+            String customDelimiter = matcher.group(1);
+            String extractedText = matcher.group(2);
+            return SplitDto.of(extractedText, customDelimiter);
+        }
+        throw new RuntimeException();
     }
-
-    static class SplitDto {
-        private String text;
-        private String delimiter;
-
-        public static SplitDto of(String text, String delimiter) {
-            SplitDto dto = new SplitDto();
-            dto.text = text;
-            dto.delimiter = delimiter;
-            return dto;
-        }
-
-        public static SplitDto of(String text) {
-            SplitDto dto = new SplitDto();
-            dto.text = text;
-            dto.delimiter = DEFAULT_DELIMITER;
-            return dto;
-        }
-
-        public String[] split() {
-            return text.split(delimiter);
-        }
-    }
-
 }
